Added support for XE3: Differences from Other Packages:
<ProjectVersion>14.4</ProjectVersion> is the correct number for XE3

- "clean" many dproj "nodes" not necessary
for example Removed VersionInfo and VersionInfoKeys nodes and "Deployment" node.

- Added "LibSuffix" 170 for XE3:
notice that LibSuffix could be added to any package from XE2 to 10_3 with fixed value. From D10_4 to D13 is recommended to use $(Auto)

<VerInfo_Locale>1033</VerInfo_Locale> into "Base" section

Changes to sources to compile with XE3:
The major chages is into UTWSVGMeasure.pas
the definition of IEUnit enumeration as public type inside TWSVGMeasure<T> is not recognized by XE3 (and probably also XE2) so it was defined as a simple enumerated type.
Addes some units to remove "Warning": now there is one Hint, and zero warnings.
changed all {$if CompilerVersion <= 23} to {$if CompilerVersion <= 24}
changed all {$if CompilerVersion > 23} to {$if CompilerVersion > 24}